Somewhat more tests incl test of Dockerfile
